smart-green-space/smart-green-space-platform/apps/ml-engine/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
tensorflow-cpu>=2.14.0
Pillow>=10.0.0
python-multipart>=0.0.6
